Fix passing options to coffee with :make (#145)
This bug was introduced by 0e38ef90446a0f1f2294b6225c9c97dbc2810b45 so that
`:make | cwindow` worked, but it broke passing in options because coffee doesn't
accept options after a filename.

Putting the $* back before the filename allows passing options but it breaks
using `:make OPTIONS | cwindow` again because vim expands that as

    coffee OPTIONS | cwindow FILE.coffee

So I updated the readme to recommend using an autocmd instead of a pipe to open
the quickfix window. This isn't perfect because it seems to be impossible to
switch to the quickfix window within an autocmd after opening it, but I'm not
sure of any other solution.